Separate ENRFactory, EndpointFactory, NodeIDFactory
