+++ ddht/tools/factories/discovery.py
@@ -1,18 +1,7 @@
-import random
-import socket
-from typing import Any, Dict, Tuple
-
-from eth_keys import keys
-from eth_utils import big_endian_to_int, int_to_big_endian
-from eth_utils.toolz import merge, reduce
 import factory
 
 from ddht.base_message import InboundMessage
-from ddht.endpoint import Endpoint
-from ddht.enr import ENR, UnsignedENR
 from ddht.identity_schemes import V4IdentityScheme
-from ddht.kademlia import compute_log_distance
-from ddht.typing import NodeID
 from ddht.v5.channel_services import InboundPacket
 from ddht.v5.constants import (
     AUTH_SCHEME_NAME,
@@ -27,7 +16,9 @@ from ddht.v5.messages import FindNodeMessage, PingMessage
 from ddht.v5.packets import AuthHeader, AuthHeaderPacket, AuthTagPacket, WhoAreYouPacket
 from ddht.v5.typing import Topic
 
-from .kademlia import AddressFactory
+from .endpoint import EndpointFactory
+from .enr import ENRFactory
+from .node_id import NodeIDFactory
 
 
 class AuthTagPacketFactory(factory.Factory):  # type: ignore
@@ -69,16 +60,6 @@ class WhoAreYouPacketFactory(factory.Factory):  # type: ignore
     enr_sequence_number = 0
 
 
-class EndpointFactory(factory.Factory):  # type: ignore
-    class Meta:
-        model = Endpoint
-
-    ip_address = factory.LazyFunction(
-        lambda: socket.inet_aton(factory.Faker("ipv4").generate({}))
-    )
-    port = factory.Faker("pyint", min_value=0, max_value=65535)
-
-
 class EndpointVoteFactory(factory.Factory):  # type: ignore
     class Meta:
         model = EndpointVote
@@ -94,95 +75,6 @@ class InboundPacketFactory(factory.Factory):  # type: ignore
 
     packet = factory.SubFactory(AuthTagPacketFactory)
     sender_endpoint = factory.SubFactory(EndpointFactory)
-
-
-class NodeIDFactory(factory.Factory):  # type: ignore
-    class Meta:
-        model = NodeID
-        inline_args = ("node_id",)
-
-    node_id = factory.Faker("binary", length=32)
-
-    @classmethod
-    def at_log_distance(cls, reference: NodeID, log_distance: int) -> NodeID:
-        num_bits = len(reference) * 8
-
-        if log_distance > num_bits:
-            raise ValueError(
-                "Log distance must not be greater than number of bits in the node id"
-            )
-        elif log_distance < 0:
-            raise ValueError("Log distance cannot be negative")
-
-        num_common_bits = num_bits - log_distance
-        flipped_bit_index = num_common_bits
-        num_random_bits = num_bits - num_common_bits - 1
-
-        reference_bits = bytes_to_bits(reference)
-
-        shared_bits = reference_bits[:num_common_bits]
-        flipped_bit = not reference_bits[flipped_bit_index]
-        random_bits = [
-            bool(random.randint(0, 1))
-            for _ in range(
-                flipped_bit_index + 1, flipped_bit_index + 1 + num_random_bits
-            )
-        ]
-
-        result_bits = tuple(list(shared_bits) + [flipped_bit] + random_bits)
-        result = NodeID(bits_to_bytes(result_bits))
-
-        assert compute_log_distance(result, reference) == log_distance
-        return result
-
-
-def bytes_to_bits(input_bytes: bytes) -> Tuple[bool, ...]:
-    num_bits = len(input_bytes) * 8
-    as_int = big_endian_to_int(input_bytes)
-    as_bits = tuple(bool(as_int & (1 << index)) for index in range(num_bits))[::-1]
-    return as_bits
-
-
-def bits_to_bytes(input_bits: Tuple[bool, ...]) -> bytes:
-    if len(input_bits) % 8 != 0:
-        raise ValueError("Number of input bits must be a multiple of 8")
-    num_bytes = len(input_bits) // 8
-
-    as_int = reduce(lambda rest, bit: rest * 2 + bit, input_bits)
-    as_bytes_unpadded = int_to_big_endian(as_int)
-    padding = b"\x00" * (num_bytes - len(as_bytes_unpadded))
-    return padding + as_bytes_unpadded
-
-
-class ENRFactory(factory.Factory):  # type: ignore
-    class Meta:
-        model = ENR
-
-    sequence_number = factory.Faker("pyint", min_value=0, max_value=100)
-    kv_pairs = factory.LazyAttribute(
-        lambda o: merge(
-            {
-                b"id": b"v4",
-                b"secp256k1": keys.PrivateKey(
-                    o.private_key
-                ).public_key.to_compressed_bytes(),
-                b"ip": o.address.ip_packed,
-                b"udp": o.address.udp_port,
-                b"tcp": o.address.tcp_port,
-            },
-            o.custom_kv_pairs,
-        )
-    )
-    signature = factory.LazyAttribute(
-        lambda o: UnsignedENR(o.sequence_number, o.kv_pairs)
-        .to_signed_enr(o.private_key)
-        .signature
-    )
-
-    class Params:
-        private_key = factory.Faker("binary", length=V4IdentityScheme.private_key_size)
-        address = factory.SubFactory(AddressFactory)
-        custom_kv_pairs: Dict[bytes, Any] = {}
 
 
 class PingMessageFactory(factory.Factory):  # type: ignore
