+++ tests/core/v5/test_message_dispatcher.py
@@ -7,11 +7,9 @@ import trio
 from ddht.base_message import InboundMessage
 from ddht.identity_schemes import default_identity_scheme_registry
 from ddht.node_db import NodeDB
-from ddht.tools.factories.discovery import (
-    EndpointFactory,
-    ENRFactory,
-    PingMessageFactory,
-)
+from ddht.tools.factories.discovery import PingMessageFactory
+from ddht.tools.factories.endpoint import EndpointFactory
+from ddht.tools.factories.enr import ENRFactory
 from ddht.tools.factories.keys import PrivateKeyFactory
 from ddht.v5.message_dispatcher import MessageDispatcher
 from ddht.v5.messages import FindNodeMessage, NodesMessage, PingMessage
