+++ tests/core/v5/test_routing_table_manager.py
@@ -13,13 +13,13 @@ from ddht.identity_schemes import default_identity_scheme_registry
 from ddht.kademlia import KademliaRoutingTable, compute_distance, compute_log_distance
 from ddht.node_db import NodeDB
 from ddht.tools.factories.discovery import (
-    EndpointFactory,
-    ENRFactory,
     FindNodeMessageFactory,
     InboundMessageFactory,
-    NodeIDFactory,
     PingMessageFactory,
 )
+from ddht.tools.factories.endpoint import EndpointFactory
+from ddht.tools.factories.enr import ENRFactory
+from ddht.tools.factories.node_id import NodeIDFactory
 from ddht.v5.constants import ROUTING_TABLE_PING_INTERVAL
 from ddht.v5.message_dispatcher import MessageDispatcher
 from ddht.v5.messages import FindNodeMessage, NodesMessage, PingMessage, PongMessage
