+++ tests/core/v5_1/test_packet_encoding.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ddht.tools.factories.discovery import ENRFactory
+from ddht.tools.factories.enr import ENRFactory
 from ddht.v5_1.messages import PingMessage
 from ddht.v5_1.packets import (
     HandshakeHeader,
